Fix Missing URL in Password Reset Emails
All emails are sent as with an HTML content-type. This breaks the
password reset URLs since Wordpress surrounds them in angle-brackets
which cause email clients to read them as invalid HTML tags. This causes
them to hide the links.

To fix this, we escape the content of the password reset emails before
sending them.