[#1051] Notify Magazine Subscribers When Downloads Are Added
When adding new downloads to the Membership Back Issues product or the
Magazine subscription, we send notifications to FIC Members &
Subscribers, letting them know that a new Download has been added & is
available for download. The new item is also attached to the
notification email.

For the Back Issues product, we also revert to the WCv2.6 behavior of
granting access to the new download for all previous orders, since all
Members should have access to new Back Issues.

Closes #930
Closes #1051