Add Searching to Directory Listings Page
Add a Search form, which applies a SearchFilter parameter to the API
request when submit.

Add a searchString field to the Model, defaulting to an empty String.

Add functions to parse the `search` query parameter & append the result
to the list of other FilterParams.

Add Routing.getSearchFilter to pull a SearchFilter out of a list of
FilterParams.

Add Routing.mapBoth to build mapPage & mapFilters.

Refactor the New Page commands to Commands.newPage.