Switch Directory Listings Page If Past Last Page
When the Pagination module receives empty items, it will ensure the
returned page is not past the maximum page number. If it is the current
page & past the total page count, the Pagination model will change to
the last page.

This is caught by the Update.update function, which then triggers the
associated URL change.