Refactor Order Creation When Joining Groups, Add Wisdom Group
Refactor the functions used for creating CMag back issues order when a
user joins the FIC Membership group. Instead of hardcoding the product &
group, a series of groups & products can be specified via the
ThemeWooCommerce::group_orders class constant.

Add handling for the Wisdom Digital group, giving users access to the
digital versions of the Wisdom of Communities volumes.