[#1335] Add UI for Configuring Flat Rate Options
Instead of hardcoding the flat-rate options in the theme, we add an
admin page allowing administrators to configure the options themselves.

The page is built using Elm and the WP REST API; the options are stored
in the WP options table and the flat rate code has been modified to pull
and use these values.

Closes #1335