[#1372] Re-Design Header Links & Icons
Redesign the login/logout, my account, & cart links in the page header.

Increase the font size, change the My Account link into an icon & move
it to the center, and space everything out a little bit.

This should make it easier for mobile users to keep from accidentally
logging out when trying to tap the other icons/links.

Closes #1372