[#1221] Add Admin Form for Editing the Board & Staff Page
This adds an Admin page & form for editing the FIC's Board & Staff page
in a more structured way, as well as Elm code for rendering the form
data on the frontend of the site.

The form presents lists of Profile forms for the Board & Staff, where
each Profile can have a name, image, and biography. The form fetches and
posts data via newly added `/board-staff/` WP-JSON endpoints. The saved
data is echo'd out in JSON as a global JS variable, which is passed to
the page rendering Elm code as flags.

Closes #1221