Return Current User Info in Details Route, Show Edit/Verify Links
determine if the current user is an admin or the listing's owner. If so,
show the appropriate links to the Edit Listing & Validate Your Listing
pages.

the validation code is actually run by a shortcode rendered on the
details page, so we need to make this into an API route.