Add Animations to Gallery Modals
Instead of relying on bootstrap's modal animation handling, we use the
elm-style-animation package to queue style changes that fade the modal
in and out, as well as fade between images when hitting the next &
previous buttons.

Also change the gallery modal to be a fixed size(dependent on screen
width), instead of changing the modal width based on the image size.