[#1393] Change Place Order Text for Memberships & Donations
When checking out with a membership or general donation, change the
Place Order button text to "Join" or "Donate" respectively. If both a
membership and donation are in the cart, use the "Donate" text. If other
products are in the cart, we do not modify the button text(for now).

Refs #1393