Update Directory Application to Elm v0.19
Restructure the URL handling by add a flag, port, and subscription for
updating the page & receiving page updates. This is required since
Wordpress still controls part of the body element, which means we can't
use the Browser.application function to manage the URL.

Modify the Model & init to fetch and store the user's timezone in
addition to the current time. This is required for showing formatted
listing dates & times.

Upgrading to v0.19 also includes lots of manual changes:

* Use the Time.Posix type instead of the removed Date.Date. This means
  we have to use new date-distance & date formatting libraries.
* Use Tuple.pair & lambdas instead of the (,) constructor.
* Use Decode.succeed instead of the removed Json.Decode.Pipeline.decode
  function.
* Change some function & argument names to prevent shadowing.
* Use the new elm/url package.
* Refactor the event attributes to deal with the removal of
  the onWithOptions function.
* Change toString calls into String.fromInt, String.fromFloat, &
  String.Conversions.fromHttpError calls.
* Refactor 4-element tuples into 3-element tuples.