[#1463] Hide Classified's "Send Message" Button When Click
When a user clicks the "Send Message" button for a classified, the
contact form slides open. Unfortunately, the form has it's own submit
button that also says "Send Message". This causes confusion, clicking
the first button will just hide the form, you need to click the second
button to actually send the message.

This fixes the confusion by hiding the first button when the form is
slid open. It also removes the duplicate "reCaptcha" header shown in the
contact form.

Closes #1463