Update Footer Content & Styling
Remove the Membership & Partners sections from the footer, adding a
Support section, 3 column mission/pages/newsletter section, & a
copyright section matching the new design.

Note that this requires adding a new mailpoet form with First/Last Name
& Email fields.

The partners section will be brought back as a shortcode on the
homepage.