Add Shortcode & Style for Get Involved/Newsletter Block
Add the `fic_get_involved` shortcode for rendering the Get Involved
sectin of the homepage with a mailpoet form. Add styles to make the
section match our mockup.

The shortcode allows customization of the header text, newsletter
header, & newsletter text. The content of the shortcode is displayed
underneath the main header.