Add SASS Styles for Backend's Block Editor
Add a new styles-editor.sass file for styles specific to the WP5 Block
Editor. For now this just sets the maximum width of the editor to the
maximum content width of the frontend.

Ensure the editor styles are not included in the frontend & add a
handler for the enqueue_block_editor_assets action to enqueue them in
the backend.