Support Full Width Sections in the Homepage
Add support for a `full-width` CSS class which allows using full-width
fluid containers inside of the main content's container.

Modify the homepage sections with colored or image backgrounds so that
they took up the full width of the page.

Remove the colored body background & drop shadow, defaulting the body's
background color to white.