Re-Style & Re-Organize the Site Header
Revamp the header according to the new design mockup.

Add spacing between the top banner & the header. Re-do the google search
code so it only shows up after clicking a search icon in the header(with
nice fade-in transitions).

Re-color the dropdown menus, limiting them to a depth of 2. There's a
caret inbetween the top-level item & the dropdown menu.

Remove the horizontal rule between the header & site content.