Add Links & Various Tweaks to Homepage Sections
Link the titles of the 3-column, Books, Events, & Offers sections on the
homepage.

Add a More Events button to the Events section since users cannot post
events.

Update the default text for the newsletter section & expand the column
width.

Link the images of the primary story sections, and the images and titles
of the secondary story sections.

Fix the alignment of the 3-column sections images & titles. Set the
title color back to black.

Some additonal styling tweaks.