Fix Nonces When Fetching Directory Details
Fix a bug causing issues for anonymous users when attempting to fetch
the details for a Community.

Now, we only add the themeRestConfig JSON object to the response when it
is being requested by a logged-in user. If the object is not present, we
simply pass an empty string to the Elm program. If the nonce is empty,
we do not include it in the GET parameters of the getCommunity command.